Add tester system for GDScript
This is meant to test the GDScript implementation (not to test user
scripts nor test the engine with scripts).

Tests are made of a folder with GDScript files containing a test()
function each and an output file (same name but with .out extension
instead of .gd).

Run Godot with "--generate-gdscript-tests path/to/tests" to generate the
.out files in the expected format. Make sure to check if the output is
the expected one.

Run Godot with "--test-gdscript path/to/tests" to run all the tests. It
will show each failed test as soon as they are found, if any.